Updated the licenses And splitted wizard layers to different services
