+++ architect-html/src/components/editors/designerEditor/common/DesignSurface.tsx
@@ -1,5 +1,5 @@
 /*
-Copyright 2005 - 2026 Advantage Solutions, s. r. o. 
+Copyright 2005 - 2025 Advantage Solutions, s. r. o. 
 
 This file is part of ORIGAM (http://www.origam.org).
