+++ backend/Origam.Architect.Server/Services/Wizards/ScreenSectionWizardService.cs
@@ -1,0 +1,68 @@
+#region license
+/*
+Copyright 2005 - 2026 Advantage Solutions, s. r. o.
+
+This file is part of ORIGAM (http://www.origam.org).
+
+ORIGAM is free software: you can redistribute it and/or modify
+it under the terms of the GNU General Public License as published by
+the Free Software Foundation, either version 3 of the License, or
+(at your option) any later version.
+
+ORIGAM is distributed in the hope that it will be useful,
+but WITHOUT ANY WARRANTY; without even the implied warranty of
+MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
+GNU General Public License for more details.
+
+You should have received a copy of the GNU General Public License
+along with ORIGAM. If not, see <http://www.gnu.org/licenses/>.
+*/
+#endregion
+
+using Origam.Architect.Server.Models.Requests.Wizards;
+using Origam.Architect.Server.Models.Responses.Wizards;
+using Origam.Schema.GuiModel;
+using Origam.Workbench.Services;
+
+namespace Origam.Architect.Server.Services.Wizards;
+
+public class ScreenSectionWizardService(
+    IPersistenceService persistenceService,
+    ModelTransactionRunner transaction,
+    SearchService searchService
+) : ScreenWizardServiceBase(persistenceService, transaction, searchService)
+{
+    public ScreenSectionWizardData GetWizardData(Guid entityId)
+    {
+        var entity = RetrieveEntity(entityId);
+
+        return new ScreenSectionWizardData
+        {
+            EntityName = entity.Name,
+            EntityCaption = entity.Caption,
+            Columns = GetScreenColumns(entity),
+            ExistingScreenSectionNames = GetScreenSectionNames(),
+        };
+    }
+
+    public CreateWizardResult CreateScreenSection(CreateScreenSectionModel input)
+    {
+        var trimmedName = RequireName(input.Name, Strings.Wizard_ScreenSectionNameRequired);
+        var entity = RetrieveEntity(input.EntityId);
+        RequireSelectedFields(input.SelectedFieldIds);
+        RequireUniqueScreenSectionName(trimmedName);
+
+        var selectedNames = RetrieveControlFieldNames(entity, input.SelectedFieldIds);
+        var groupName = entity.Group?.Name;
+
+        var panel = Transaction.Run(() =>
+        {
+            var newPanel = GuiHelper.CreatePanel(groupName, entity, selectedNames, trimmedName);
+            RelayoutFields(newPanel);
+            SetPanelTitle(newPanel, input.Caption);
+            return newPanel;
+        });
+
+        return BuildResult([panel]);
+    }
+}
